refactor de componente
